dbs/postgres-db/pyproject.toml
[tool.poetry]
name = "postgres-db"
version = "0.1.0"
description = ""
authors = ["Antoine Gerardin <antoine.gerardin@nih.gov>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
polus-aithena-common = {path= "../../common", develop=true}
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
